api/requirements.txt
coverage==4.2
decorator==4.0.10
Django==1.10.1
django-cors-headers==2.0.2
django-debug-toolbar==1.5
django-extensions==1.7.3
django-filter==0.13.0
djangorestframework==3.4.6
djoser==0.5.0
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
Markdown==2.6.6
path.py==8.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
ptyprocess==0.5.1
Pygments==2.1.3
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.1
traitlets==4.2.2
wcwidth==0.1.7
Werkzeug==0.11.10
